Move sync modal
