--- frontend/src/metabase/containers/SyncingModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SyncingModal";
