+++ frontend/src/metabase/databases/components/DatabaseSyncModal/DatabaseSyncModal.styled.tsx
