+++ frontend/src/metabase/databases/components/DatabaseSyncModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DatabaseSyncModal";
