+++ frontend/src/metabase/databases/containers/DatabaseSyncModal/DatabaseSyncModal.tsx
@@ -1,11 +1,12 @@
 import { connect } from "react-redux";
 import { createSelector } from "reselect";
 import _ from "underscore";
-import Settings from "metabase/lib/settings";
 import Databases from "metabase/entities/databases";
 import DatabaseCandidates from "metabase/entities/database-candidates";
-import SyncingModal from "metabase/components/SyncingModal";
+import { getSetting } from "metabase/selectors/settings";
 import { Database, DatabaseCandidate } from "metabase-types/api";
+import { State } from "metabase-types/store";
+import DatabaseSyncModal from "../../components/DatabaseSyncModal";
 
 interface DatabaseProps {
   databases: Database[];
@@ -16,10 +17,10 @@ interface CandidatesProps {
 }
 
 const getSampleQuery = createSelector(
-  (state: unknown, props: DatabaseProps) => props.databases,
-  databases => {
+  (state: State, props: DatabaseProps) => props.databases,
+  (state: State) => getSetting(state, "enable-xrays"),
+  (databases, enableXrays) => {
     const sampleDatabase = databases.find(d => d.is_sample);
-    const enableXrays = Settings.get("enable-xrays");
 
     if (sampleDatabase && enableXrays) {
       return { id: sampleDatabase.id };
@@ -44,4 +45,4 @@ export default _.compose(
   Databases.loadList(),
   DatabaseCandidates.loadList({ query: getSampleQuery }),
   connect(mapStateToProps),
-)(SyncingModal);
+)(DatabaseSyncModal);
