+++ frontend/src/metabase/databases/containers/DatabaseSyncModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DatabaseSyncModal";
